ICU end-of-life PTSD survey data

Data from a 2021–2022 study of family members of ICU patients who were isolated from loved ones during the COVID-19 pandemic. Respondents were assessed for PTSD using the Impact of Events Scale-6 (IES-6); a score of 10 or higher indicates likely PTSD.

covid_eol

Format

A data frame with 298 rows and 16 variables:

subject_id

Integer. Unique subject identifier.

ies_total

Integer. IES-6 total score (0–24).

age

Integer. Respondent age in years.

ethnicity

Factor with levels Hispanic, Non-Hispanic.

race

Factor with levels AAPI/NA/NI, Black, Other, White.

sex

Factor with levels Female, Male.

education

Factor with levels <= High school diploma or GED, 4-year college degree, Some college or trade school, Some graduate school or graduate degree.

relationship

Factor with levels Child, Other, Parent, Sibling, Spouse/partner. Relationship of the respondent to the ICU patient.

resp_failure

Factor with levels High Flow nasal cannula, Invasive MV, nasal cannula, Non-invasive MV. Type of respiratory failure treatment.

ecmo

Factor with levels No, Yes. Whether extracorporeal membrane oxygenation was used.

icu_totaldays

Integer. Total days in the ICU.

death

Logical. Whether the ICU patient died.

psych_meds

Factor with levels No, Yes. Whether the respondent was taking psychiatric medications.

charlson

Integer. Charlson Comorbidity Index score.

sofa

Integer. Sequential Organ Failure Assessment score (0–20).

ICU_admit_month

Factor with levels April, Feburary/March, May/June/July. Month of ICU admission.

Source

De-identified data from an ICU family-member survey; used with permission.